Add files via upload
This project focuses on the harmful effects of pollution caused by vehicle emissions, industrial activities, and removing trees on the environment and human health. It highlights how these pollutants negatively impact people of all ages, from children to the elderly, contributing to severe health problems like respiratory illnesses, cardiovascular diseases, and weakened immunity. The project emphasizes the urgent need for sustainable solutions to protect the environment and safeguard the health of current and future generations.